+++ .gitignore
@@ -10,4 +10,3 @@ node_modules/
 npm-debug.log*
 Thumbs.db
 .vscode/
-functions.js
